+++ src/cli.rs
@@ -6,9 +6,8 @@ use std::{
 
 use anyhow::Result;
 use clap::{Parser, Subcommand};
+use clap_complete::Shell;
 use miette::{LabeledSpan, MietteDiagnostic, NamedSource};
-use serde::{Deserialize, Serialize};
-use serde_with::skip_serializing_none;
 
 use crate::{
     config::{Config, Req},
@@ -22,15 +21,24 @@ const VERSION: &str = env!("CARGO_PKG_VERSION");
 #[cfg(feature = "solar")]
 const VERSION: &str = concat!(env!("CARGO_PKG_VERSION"), "-solar");
 
-#[derive(Subcommand, Debug, Clone, Copy, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
+#[derive(Subcommand, Debug, Clone)]
 pub enum Commands {
     /// Create a `.lintspec.toml` config file with default values
     Init,
+
+    /// Generate shell completion scripts
+    Completions {
+        /// The flavor of shell for which to generate the completion script
+        #[arg(short, long)]
+        shell: Shell,
+
+        /// The output directory for the file, prints to `stdout` if omitted
+        #[arg(short, long, value_hint = clap::ValueHint::DirPath)]
+        out: Option<PathBuf>,
+    },
 }
 
-#[derive(Parser, Debug, Clone, Serialize, Deserialize)]
-#[skip_serializing_none]
+#[derive(Parser, Debug, Clone)]
 #[command(version = VERSION, about, long_about = None)]
 #[non_exhaustive]
 pub struct Args {
@@ -236,7 +244,7 @@ macro_rules! cli_rule_override {
 }
 
 /// Read the configuration from config file, environment variables and parsed CLI arguments (passed as argument)
-pub fn read_config(args: Args) -> std::result::Result<Config, figment::Error> {
+pub fn read_config(args: Args) -> Result<Config, Box<figment::Error>> {
     let config_path = args
         .config
         .or_else(|| env::var("LS_CONFIG_PATH").ok().map(Into::into));
@@ -314,7 +322,7 @@ pub fn print_reports(
     file_diags: FileDiagnostics,
     contents: String,
     compact: bool,
-) -> std::result::Result<(), io::Error> {
+) -> Result<(), io::Error> {
     if compact {
         for item_diags in file_diags.items {
             item_diags.print_compact(f, &file_diags.path, &root_path)?;
@@ -339,7 +347,7 @@ fn print_report(
     f: &mut impl io::Write,
     source: Arc<NamedSource<String>>,
     item: ItemDiagnostics,
-) -> std::result::Result<(), io::Error> {
+) -> Result<(), io::Error> {
     let msg = if let Some(parent) = &item.parent {
         format!("{} {}.{}", item.item_type, parent, item.name)
     } else {